requirements.txt
streamlit
pandas
plotly
google-generativeai
vnstock
beautifulsoup4
ipython